Merged PR 99: Better email handling - email is sent centrally from its own console app
Rip out old email code and replace with an email handler script designed to be run on a schedule from a central place on the server side. This prevents denial of service attacks (intentional or unintentional) from clients who spam buttons (though it will still create a ton of orders if they do this).